Large change to TO calculation.
TO is calculated on each pixel but we do not need a 64 float array of
v_ir_off_comp and v_ir_tgc_comp. Instead, we can use one float variable
that gets calculated each time around the TO loop. This saves ~256 bytes
of RAM and makes the code slightly easier to read as it reflects the
datasheet more directly.

There were a few other whitespace and variable name changes for
readability.